Revert "Move key_management schemas and modules into central datastore module"
This reverts commit 04c1f425dbd7b7856b5e52acc53ce0417206c154.
