+++ libsplinter/src/biome/key_management/store/diesel/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub(in crate::biome) mod models;
 mod operations;
 #[cfg(feature = "postgres")]
 pub(in super::super) mod postgres;
+mod schema;
